requirements.txt
Flask>=3.0.0,<4.0.0
Flask-Login>=0.6.3,<1.0.0
Flask-SQLAlchemy>=3.1.0,<4.0.0
PlexAPI>=4.15.0,<5.0.0
python-dotenv>=1.0.0,<2.0.0
Flask-Limiter>=3.5.0,<4.0.0
Flask-WTF>=1.2.0,<2.0.0
gunicorn>=21.2.0,<22.0.0
psycopg2-binary>=2.9.9,<3.0.0
SQLAlchemy>=2.0.0,<3.0.0

